requirement.txt
streamlit
openai>=1.0.0
azure-storage-blob
pymupdf